added a bit of modification from the routes, added pagination

+++ database/migrations/2023_12_03_134147_create_questions_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('questions', function (Blueprint $table) {
             $table->id();
             $table->text('question_text');
-            $table->unsignedInteger('correct_answer');
+            $table->unsignedBigInteger('correct_answer');
             $table->foreignId('user_id')->constrained()->onUpdate('cascade')->onDelete('cascade');
             $table->timestamps();
         });
